Tidy tests
Reduce the amount of code in individual files and make switching
between source and tests easier.
